pyproject.toml
[tool.poetry]
name = "AWESOME-FASTAPI-PROJECT"
version = "0.1.0"
description = ""
authors = [
    "Yaroslavshym <139383618+Yaroslavshym@users.noreply.github.com>",
    "Vasyl Kartychak https://github.com/smartguy-coder/AWESOME-FASTAPI-PROJECT",
    "Illya Kartychak",
    "Kiril Lypenko",
    "Danylo Soldatenko <dansoldat9@gmail.com>",
    "Konstantin Nezhalsky"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
fastapi-versioning = "^0.10.0"
sentry-sdk = "^1.35.0"
pymongo = "^4.6.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
fastapi-mail = "^1.4.1"



[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mkdocs-material = {extras = ["imaging"], version = "^9.4.14"}
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
faker = "^21.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.pytest.ini_options]
asyncio_mode = "auto"


[tool.logging.loggers.project.passlib]
level = "ERROR"
